break-out NoResultsDisplay into dedicated component, shared by Catolog and Compare
