+++ src/components/CatalogView/index.js
@@ -4,32 +4,12 @@ import { useRecoilValue } from 'recoil';
 import './CatalogView.css';
 import SelectedCardCollection from '../SelectedCardCollection';
 import RecordSelector from '../SelectedCardCollection/RecordSelector';
+import NoResultsDisplay from '../NoResultsDisplay';
 import {
   activeCategoriesState,
   activeProvidersState,
   filteredActiveCollectionState,
 } from '../../recoil';
-
-const noneEnabled = (obj) => Object.values(obj).reduce((acc, isEnabled) => (isEnabled ? false : acc), true);
-
-const NoResultsDisplay = React.memo(({ filteredRecordCount, activeCategories, activeProviders }) => {
-  if (filteredRecordCount) {
-    return null;
-  }
-
-  let message = 'No data found for the selected Records, Providers, and Time period';
-  if (noneEnabled(activeCategories)) {
-    message = 'No Record type is selected';
-  } else if (noneEnabled(activeProviders)) {
-    message = 'No Provider is selected';
-  }
-
-  return (
-    <div className="tiles-view-container-inner-empty" key="1">
-      { message }
-    </div>
-  );
-});
 
 const CompareView = () => {
   const filteredActiveCollection = useRecoilValue(filteredActiveCollectionState);
