+++ src/components/CompareView/index.js
@@ -4,6 +4,7 @@ import { useRecoilValue } from 'recoil';
 import './CompareView.css';
 import SelectedCardCollection from '../SelectedCardCollection';
 import RecordSelector from '../SelectedCardCollection/RecordSelector';
+import NoResultsDisplay from '../NoResultsDisplay';
 import Sparkline from '../Sparkline';
 
 import {
@@ -22,27 +23,6 @@ const formatYearRange = (minDate, maxDate, pre, post) => {
 
   return minYear === maxYear ? pre + minYear + post : `${pre + minYear} \u2013 ${maxYear}${post}`;
 };
-
-const noneEnabled = (obj) => Object.values(obj).reduce((acc, isEnabled) => (isEnabled ? false : acc), true);
-
-const NoResultsDisplay = React.memo(({ filteredRecordCount, activeCategories, activeProviders }) => {
-  if (filteredRecordCount) {
-    return null;
-  }
-
-  let message = 'No data found for the selected Records, Providers, and Time period';
-  if (noneEnabled(activeCategories)) {
-    message = 'No Record type is selected';
-  } else if (noneEnabled(activeProviders)) {
-    message = 'No Provider is selected';
-  }
-
-  return (
-    <div className="tiles-view-container-inner-empty" key="1">
-      { message }
-    </div>
-  );
-});
 
 const ProviderSparkLine = ({
   providerLabel, uuids, records, minDate, maxDate,
